src/lambdas/api_endpoints/requirements.txt
annotated-types==0.7.0
aws-lambda-powertools==3.8.0
jmespath==1.0.1
psycopg2-binary==2.9.10
pydantic==2.10.6
pydantic-core==2.27.2
pydantic-settings==2.8.1
python-dotenv==1.0.1
typing-extensions==4.12.2
